flats_project/requirements.txt
Scrapy==2.11.0
flask==3.0.1
sqlalchemy==2.0.25
python-dotenv==1.0.1
psycopg2